requirements.txt
docx2pdf==0.1.7
fpdf==1.7.2
img2pdf==0.3.6
Pillow==7.2.0
PyPDF2==1.26.0
pywin32==227
tqdm==4.48.0
